And again a better starting point, I just had to put in the red labels so you (and I) could figure out why this program is useful.
A Contest Manager (Roy) asked me to create an SUA for a Finish Line at Harris Hill so I made this program.

A good example would be to say, for Perry, I want an SUA for a 1 km long finish line which goes from the fire pit, perpendicular to the runway.  Use Google Earth and its 'draw line' feature to get the end point and direction, then fill that into program here.